ollama_examples/pyproject.toml
[project]
name = "ollama_examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "econagents[all]>=0.0.3",
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
]
